services/hyperagent-tools/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0
py-solc-x>=2.0.0
